only depuplicate item with a subjext

+++ deduplicate-calendar.py
@@ -7,6 +7,7 @@ def opt_args():
     parser = kopano.parser('skpcUPm')
     parser.add_option('--user', dest='user', action='store', help='username')
     parser.add_option('--folder', dest='folder', action='store', help='foldername')
+    parser.add_option('--subject', dest='subject', action='store', help='subject')
     return parser.parse_args()
 
 def main():
@@ -20,21 +21,22 @@ def main():
     read_items = []
     num = 0
     for item in folder.items():
-        md5 = hashlib.md5(item.prop(PR_SUBJECT).value)
-        if item.get_prop(PR_START_DATE):
-            md5.update(item.prop(PR_START_DATE).value.strftime(("%m/%d/%Y, %H:%M:%S")).encode())
-        if item.get_prop(PR_END_DATE):
-            md5.update(item.prop(PR_END_DATE).value.strftime(("%m/%d/%Y, %H:%M:%S")).encode())
-        if item.get_prop(PR_BODY):    
-            md5.update(item.prop(PR_BODY).value)
-        if item.get_prop(PR_HTML):    
-            md5.update(item.prop(PR_HTML).value)
-        md5 = md5.hexdigest()
-        if md5 in read_items: 
-            num += 1
-            folder.delete(item)
-        else:
-            read_items.append(md5)
+        if item.prop(PR_SUBJECT).value == options.subject.encode():
+            md5 = hashlib.md5(item.prop(PR_SUBJECT).value)
+            if item.get_prop(PR_START_DATE):
+                md5.update(item.prop(PR_START_DATE).value.strftime(("%m/%d/%Y, %H:%M:%S")).encode())
+            if item.get_prop(PR_END_DATE):
+                md5.update(item.prop(PR_END_DATE).value.strftime(("%m/%d/%Y, %H:%M:%S")).encode())
+            if item.get_prop(PR_BODY):    
+                md5.update(item.prop(PR_BODY).value)
+            if item.get_prop(PR_HTML):    
+                md5.update(item.prop(PR_HTML).value)
+            md5 = md5.hexdigest()
+            if md5 in read_items: 
+                num += 1 
+                folder.delete(item)
+            else:
+                read_items.append(md5)
 
     print('deleted {} items in folder {}'.format(num, folder.name))
